Nest TickContainer in BeatSnapVisualiser
